+++ History.md
@@ -1,6 +1,7 @@
 0.1.0 - July 15, 2013
 ---------------------
 * changed to factory from constructor
+* added `construct` event
 
 0.0.1 - June 9, 2013
 --------------------